requirements.txt
beautifulsoup4==4.11.1
ddddocr==1.4.3
logzero==1.7.0
numpy==1.22.3
Pillow==9.1.1
PyExecJS==1.5.1
requests==2.23.0
SocksiPy_branch==1.01
